CIAO/lib_tezos_ci: a job may not declare an [image]
We are adapting CIAO to be able to express a [container_scanning] job for a specific Docker image.
Its job definition will refer to no image (because an image is already defined in the template).

Thus we need to modify the [job] definition in CIAO so that providing an image is optional instead of mandatory.