RISC-V: support only NonZeroXRegister in ICB
Not having to handle `x0` simplifies the logic for dealing with this register. It is the goal of the 'OpCode Refinement' project to remove `x0` entirely, from the machine_state instructions. 

Therefore, we make use of this and declare an instruction is only 'ready to be lowered' once `NonZeroXRegister` is properly used.