EVM Node: Fix `eth_getLogs` refusing `"topics" : null`
Partners have reported the EVM node wrongly returns an error on `eth_getLogs`
requests with `topics` set to `null`. After a quick investigation on a public
node provider for EVM chains (Quicknode), it appears that `null` is a valid
input, and behaves like if `topics` is omitted.

We implement this behavior for this specific case. We may want to redefined
`opt` as `dft ... None` globally, if the issue arises again.