EVM Node: Do not fetch already known blocks when replaying blocks
The new block storage is incompatible with the early blocks of Etherlink
Mainnet and Testnet. To circumvent this limitation, we have introduced a
dedicated experimental feature which allows an EVM node in observer mode to
fetch the missing information from its upstream node. This was done with one
use case in mind: bootstrapping a chain with a node using the new storage. But
there are legitimate use cases for replaying blocks that we already know, and
it is a waste to require the node to fetch data it already has available in its
local SQLite database.

In order to avoid a needless request, we extend the code to first check locally
if the block is available, before trying to fetch it online. The nice benefit
of this approach is that it enables supporting the feature flag in sequencer
mode (which is a good first step towards removing the feature flag when we
stabilize it).