EVM Node: Stabilize the new block storage
The EVM node has an experimental features allowing to store blocks in the
SQLite store, instead of keeping them in the Irmin context. Our experiments are
showing that doing so results in a net win wrt. data directory size and GC
duration.

As a consequence, we stabilize the experimental feature.