CI: Do not use `npx` when running in the CI
`npx` by default always make a GET request to the npm registry, for some
reasons. Even with `--no-install`. Sadly, with the current CI setup, this call
is obvservably flaky, as demonstrated by the current Tezt job failures.

To avoid using `npx`, we rely on the fact `eth-cli` and `solc` are installed
globally in the Docker image, and hardcode the path to the installed binaries.