EVM Kernel: Use the timestamp from IPL in sequencer mode
Following the incident of Ghostnet, Etherlink Testnet experienced a downtime
which was the result of a flush of the delayed inbox. Interestingly, this flush
has occurred after 10h, not 12h as expected.

This is is because in the kernel today, we estimate the timestamp to use to
determine in an item is overdue by using the timestamp coming from the L1
(which is the timestamp of the previous block) to which we add an approximation
of the time between block.

This works well when the L1 is healthy. Not that much when there is a 2-day
downtime which greatly increases the estimated downtime.