EVM Node: Drop default rolling and store gc parameters in the store
In this patch, we force users to explicitely set the desired retention period
when they use the rolling mode, and we store this information in the store.
This forces users to decide how much history they will keep, which will reduce
the risk for us to introduce breaking changes by modifying the default
configuration.

This patch also remove the ability to switch from `rolling` to `archive`. This
could be introduced later, but at first, it is probably less confusing to force
a user to reimport an archive node instead.