EVM Node: Reconstruct transaction objects with sqlite3 backend
The transaction object stored by the kernel in its durable storage is a lie.
More precisely, it is a faitfull transaction object only for _legacy_
transactions. For more recent formats that we support like EIP 1559, some key
information are lost as the real object is projected into a legacy object.

This can be a real issue, as these lost data can be useful. For instance, our
EIP-1559 transactions are not displayed as such in Blockscout. Besides, this
can break some general-purpose tools, as our partners have found out.

In this patch, we propose to reconstruct the original, correct transaction
object from the one stored by the kernel, using the blueprints containing the
originar raw transactions. This has an impact on RPCs performance, and we
should improve the situation by storing the correct information in the store.
But this first proposal has the significant advantage to be transparent for
node providers, and to fix the issue for users.

Note that there is some redundancy between the new `Transaction_object` module
introduced in this MR and the transaction-related modules used to do the native
verification of transactions. Factoring the code for these two use cases is
probably possible, and should definitely be done in a follow-up.