P2P: refactor worker_loop of [P2p_maintenance]
This is peneloping chapter 1. This refactoring is meant to facilitate
the port to tezos-workers. I believe the logic is equivalent in this
refactoring, and if not, the following commit porting to tezos-workers
will be invalid.

The legacy worker_loop does:
1. If it needs to be maintained, maintain and loop
2. If not, send swap request and wait for maintenance motive

Therefore, I created a recursive maintain that
1. If it needs to be maintained, maintain and loop
2. If not, send swap request and return

The point to do this is that the new recursive function `maintain` is
the core logic, whereas worker_loop is responsible to wait for a
motive. We therefore just need to port worker_loop and it's
transparent that maintenance logic does not change.