Tezt cloud/Etherlink+DAL: configure sequencer catchup mechanism
This patches the parameters related to the catchup mechanism.

max_blueprint_lag is how much lag (in L2 levels) the sequencer
tolerates before resending blueprints on the
inbox. max_blueprint_ahead is how much lag it tolerates before locking
the tx-pool.

Setting max_blueprint_lag to 300 makes the sequencer trigger catchup
fast when a blueprint is actually missing (typically, because the DAL
slot containing it was not attested) while tolerating a bit of lag
from the rollup node.

For max_blueprints_ahead, 2000 is a very high value, much higher than
what we typically see for the confirmation lag in the experiments.