EVM Node: Fix switching from `rolling:n` to `rolling:m` when m < n
When trying to switch to a new, shorter retention period, the node does not
behave as expected. Indeed, the current implementation simply clean-up the
latest day of history, not accounting for the fact that we now want to keep
less history.

The present patch amends the logic of the GC to fix this oversight, by
selecting the first chunk too old instead of the oldest one.