EVM Node: Fix RPC nodes unable to start with full contexts
The latest release of the EVM node does not allow to start a RPC process using
a full context. This is because the `Evm_ro_context.preload_known_kernels`
function used by the RPC processes relies on the kernel upgrade events stored
in the SQLite db to know which Evm states to use to preload kernel. The full
mode keeps these events (because they are necessary to stream blueprints), but
it does not keep the associated states. As a consequence, a RPC process will
exit with an error at runtime, when trying to preload a kernel from a pruned
state.

We fix this bug by aligning the implementation with the `Evm_context` module:
when a state expected to be used to preload is missing, instead of failing, we
just skip it.