EVM Node: Hygienic `snapshot export`
Similarly to what was done before for `snapshot import`, we make sure
`snapshot export` is hygienic. That means that,

  1. The command creates temporary files in its data directory
  2. The command cleans up correctly its temporary files if there is an error
     or it is interrupted.

Sadly, for now it means no more fancy progress bars for us. Instead, we
periodically report progress, just like when we download a snapshot for
instance.

Here is an example of the new output.

    Feb 26 15:47:21.867: exporting snapshot snapshot.gz
    Feb 26 15:47:35.093: compressing snapshot snapshot.gz
    Feb 26 15:48:35.095: still compressing snapshot.gz after 1min (27.9% done)
    Feb 26 15:49:35.115: still compressing snapshot.gz after 2min (61.2% done)
    Feb 26 15:50:35.130: still compressing snapshot.gz after 3min (87.8% done)