RISCV: Implement try_from for `tkill` parameters
Adds types for each of `tkill`'s parameters and implements functionality
to implement conversion into these types after reading from registers.
Inputs which are out of bounds result in failure to create the types.

Implements `from` for TryFromIntError to allow for chaining the question
mark operator to bubble-up errors when inputs are not within a valid
range.