EVM Node: Stricter parsing of L2 address coming from CLI arguments
The previous implementation for existing commands like `make sequencer upgrade
command` or `make kernel installer config` was very light. The only check was
to ensure the string passed as an argument was prefixed by `0x`.

To solve this, we re-use the parsing function introduced for the `--fund`
argument to pre-existing commands.