Etherlink/Kernel: introduce read_current_blueprint_header
During stage 1, we don't care about the chain-specific part of block
headers; we only care about the part called the blueprint header.

This commit defines in blueprint_storage a new function which reads
from the storage the current block header and only exposes the
blueprint header. It also makes use of this new function instead of
read_current_block_header when possible.