EVM Kernel: Introduce regressions for events from precompiled contracts
Events are part of our user-facing APIs. Our users will parse our events to
know what is happening on-chain. From this perspective, is it very important to
know when we break our encodings, so that we can properly communicate about it
(or better still, not break them at all).

This patch adds regression scenarios that will make it obvious when we break
the events emitted by the kernel (on deposits and withdrawals).