EVM/Tezlink/Node: Get irmin hash takes a chain family to know how it should decode the block
This commit parametrize get_irmin_hash_from_block_hash but the call to get_irmin_hash force
the node to think it's in etherlink mode. The variant branch Block_hash where get_irmin_hash
is call is a parameter specific to ethereum so we are naturally in EVM mode.
For tezlink we'll call get_irmin_hash in another function.