Tezos_clic: Allow binaries to set the formatter margin
With the output of `man` being correctly wrapped, it becomes comfortable to see
`Tezos_clic` adjusts the `Format` to take into account the number of columns
available.

While it can be appealing to hide this behind `Tezos_clic`’s API, the fact that
this library takes a `Format.formatter` makes that questionable. It is
noteworthy that the other terminal-related task (deciding if the format is
`plain` or `ansi` based on the nature of `stdout` and `stderr`) is delegated to
callers, not `Tezos_clic` itself.

This is why, in this commit, we propose to add a new argument to
`setup_formatter`, effectively delegating the choice of the number of columns
per line to use to print the manual to the outside world.