EVM Node: Default data directory printed with $HOME
When using `octez-evm-node man run observer -v 3` for instance, the default
values for arguments supporting them is printed. The issue is that, when the
default argument is computed at startup using environment variable, the
computed version is printed, not the “original formula.”

To give a proper example, without this MR, `man` prints

    Start the EVM node in observer mode
    --data-dir <data-dir>: The path to the EVM node data directory
      If set, defaults to the value of EVM_NODE_DATA_DIR, or
      `/home/lthms/.octez-evm-node` otherwise.

while with this MR, `man` prints

    Start the EVM node in observer mode
    --data-dir <data-dir>: The path to the EVM node data directory
      If set, defaults to the value of EVM_NODE_DATA_DIR, or
      `$HOME/.octez-evm-node` otherwise.

My claim is that the latter is more useful than the former.