Etherlink/Tests: fix a conflict between bootstrap1 operations
In test_store_smart_rollup_address, the bootstrap1 account was used
both by the smart-rollup node to publish inbox messages (the blueprint
chunks from the sequencer) and by the client to originate another
rollup.

When both happen at the same L1 level we get conflicting operations in
the mempool.

We fix this by using another bootstrap account to originate the rollup.