Tezt: Improve the Fast Withdrawal L1 contract to record payout target
The contract in `master` does not record the target of the payout submitted by
a payer. This means a payer can provide their funds to themselves, and keep
being paid in the end. It means it does not allow us to test scenarios where
the payer lies, or when there is bug in the precompile which leads to sending
the wrong target.

This amended L1 contract hopefully does not suffer the same shortcomings, and
is a better candidate for extensive testing of the feature.

Co-authored-by: Julien Coolen <julien.coolen@nomadic-labs.com>