EVM Node: Revert to using mixHash
In a7a8437, we have modified the JSON encoding of the block to use `prevRandao`
instead of `mixHash`. This was the first of a series of changes required to
properly support [EIP-4399]. It turns out that, despite EIP-4399 being live
for almost 2 years, public RPC endpoints keep using `mixHash` in their JSON
encodings for blocks.

Notably,

  - https://arb1.arbitrum.io/rpc
  - https://mainnet.optimism.io/

both return `mixHash` (as well as the Ethereum RPC endpoint I’ve tested, and the
example endpoints of main RPC node providers like Infura, Quicknode, etc.

For these reasons, we suggest reverting the initial change, and properly
implement EIP-4399 starting with the next kernel upgrade proposal.

[EIP-4399]: https://eips.ethereum.org/EIPS/eip-4399