Images/CI: remove [VERSION]
This file was added as a way to force the rebuild of CI Docker images:
modifying it would modify the hash computed in `ci_create_ci_images.sh` which would trigger the rebuild of CI images.

Since the merge of https://gitlab.com/tezos/tezos/-/merge_requests/17055, we have a more powerful and robust mechanism.
Hence this file is not needed anymore.