Etherlink/Kernel: extract chain config from the config record
We want to allow having 3 possible types for the chain
config (EvmChainConfig, MichelsonChainConfig, and an enum wrapping
both cases simply named ChainConfig). Updating a field in a record to
change its type is syntactically too heavy compared to simply passing
the chain config alongside the config when we need both.