Clic: consider - as a valid parameter/argument value
We now consider arguments of the form `--arg -` as being of arity 1 (where `-`
is the value), whereas before we would consider `--arg` as arity 0 and then try
to parse `-` as another argument.

There is no way to define an argument of the form `-` as short arguments are
always a single character like `-p`. So here we only consider the argument a
switch if the next one is at least 2 characters long and starts with `-`.