Etherlink/Tests: in test_clean_bps, send BP 42 after BP 0
This commit adds a bake_until_sync at the beginning of the test
entitled "All blueprints are cleared on flush" so that the first
blueprint produced in the test (which has number 42) is not produced
too early. In particular, we want it to be produced strictly later
than the genesis blueprint produced by the sequencer.