DAL/Tezt: remove obsolete test disabled due to GossipSub scoring change
This test was previously disabled following changes to the GossipSub
scoring mechanism, which we partially disabled to better align with
the DAL design. In the DAL context, attesters primarily receive data
and rarely send it, which causes them to accumulate poor scores and
be disconnected by peers.

Since there is currently no plan to re-enable this part of the scoring,
the test is being removed in this MR.