EVM node: fallback once on 401 and 403 for EVM events batch fetching
The rollup node RPC durable/wasm_2_0_0/values can be unreachable even
on up-to-date rollup nodes because of ACL filtering or external HTTP
webserver rules. This makes the EVM node consider these errors as
permanent instead of transient and prevents the EVM node from repeatedly
trying to call it and instead fallback to fetching events one-by-one.