EVM Node: Refactor blueprints follower
This patch is the first step towards improving the performances of the
`Blueprints_follower` module. The current approach of this module is
particularly naive: on catch-up, each blueprint is individually fetched using a
new HTTP connection. This is very inefficient, and it shows for nodes that are
located far from the sequencer endpoint.

To improve the situation, we first refactor the module to cleanly split
blueprints fetching and processing, by introducing the `Blueprints_sequence`
module. This module relies on the `Lwt_seq.t` type to create a lazy sequence of
blueprints that the `Blueprints_follower` module can fold over.

This will make fetching more than one blueprint at a time a lot more easier, as
it is simply a different way to construct the sequence.

Co-authored-by: Raphaël Cauderlier <raphael.cauderlier@nomadic-labs.com>