EVM Node: Add `GET /evm/blueprints/range` to fetch several blueprints at once
The implementation is probably naive, but the bottleneck we are trying to
remove is not SQLite latency. It’s network latency. Tested locally, fetching
500 blueprints with this RPC locally is almost instant (around 10ms).

We also add a hard-coded upper bound on the number of blueprints one can fetch:
500. This means we will be able to leverage this RPC to divide by 500 the
number of HTTP requests a node has to do to bootstrap, while limiting the risk
of this RPC being used to take down a node.