EVM Node: Recover from inconsistent streamed RPC responses
The blueprint follower does not try to verify the consistency of what the
remote EVM node is sending. This task is delegated to the callers. Both the
`Observer` and `Rpc` modules were failing to do the proper check. They were
returning `Continue` in places where the number of the blueprint was
inconsistent.

This patch addresses that, by making sure we only return `Continue` when
everything went fine with a blueprint of a given number. We also rename
`Check_for_reorg` into `Restart_from` as it describes better the goal of this
variant.