RISC-V: remove hashing from interpreted block
Interpreted mode has no need to hash the blocks - this was done initially to avoid complications re block metrics (which do required the hash).

Hashing is now done in the BlockMetrics wrapper and JIT only (it is required for JIT, to give a unique name to blocks).  In future, hashing will move into the 'JitCompilationDispatcher', as part of the OfflineJIT work