EVM Node: Add `--replicate` to `run sandbox`
This argument can be used to have a sandbox endpoint shadowing an existing
chain locally. This can be useful for a number of use cases, notably
interacting with the chain in a sandbox environment that reflects changes
happening on the network in real time, or evaluating the impact of a (future)
kernel on realistic traffic (using `--kernel`).

This was already useful to assert that D indeed address the gas price spikes we
are seeing on Mainnet.