EVM Node: Recover from a bad gateway error
The EVM node is supposedly resilient to HTTP errors from the upstream nodes it
is connected to. Unfortunately, there is one error that it does not handle
well: 500 (Bad Gateway).

This is because the Bad Gateway error is handled in an opinionated way. Namely,
when a Bade Gateway error is returned, the HTTP client library will retry 10
times, before giving up. See `RetryClient.call_with_closefn` (in
`RPC_client_unix.ml`). After 10 attemps, the Bad Gateway error is returned
as-is. When this happens, `generic_call_loop` in `client.ml` (Resto) wraps the
error in `Unexpected_status_code`, which is not considered a connection error
by our retry logic. For good measure, we also guard against gateway timeout and
unavailable service error.

This patch addresses this issue, and should make the EVM node resilient to bad
gateway errors. This patch highlights how cumbersome and inconvenient our HTTP
stack is when it comes to make error handling. There is two levels of
indirection, when it come to errors.