EVM Node: Exit with an error when on blueprints follower failures
This should never happen. If it does, we could either retry, or simply exit
with an error. Exiting with an error provide the opportunity to cleanly restart
the process, which has proven to be an effective approach to restore node
functionality.

We have seen both observer and RPC processes becoming partially stalled, and
considering they both use the `let* _ and* _ in` pattern, it would make sense.
Indeed, if the blueprints follower returns an error, this error will never been
inspected because the other never returning promises will never been fulfilled.