Etherlink/Node/Tx_queue: propagate the Name module
This commit is needed to avoid registration conflict when the
Tx_container module will be instantiated several times (once for
Etherlink and once for Tezlink).

Each instantiation of the Tx_container functor leads to the
registration of the tx_queue_is_closed error and some events. They
need to have a different id each time they are registered, which we
can achieve by using different `Name` modules (and in particular
different `Name.base` lists).