EVM Node: Make `eth_gasPrice` safer to use
We want `eth_gasPrice` to give a result that makes it very unlikely that the
transaction is refused by the sequencer because of a gas price variation. As a
consequence, we will over-approximate the variation of the gas price by
computing the next gas price with these two inputs:

  – Arbitrary double the backlog
  – As if the block of inclusion is done in the very same second as the latest
    block

This should lead `eth_gasPrice` to provide an over-approximation when the gas
price is not the default one, but still keep the output of `eth_gasPrice` to
the minimal one when the chain is not overloaded.

In order to avoid approximating too much, we cap the gas price increase to 33%
of what it was before. For a current gas price of 1.5Gwei, the result of
`eth_gasPrice` cannot be larger than 2 Gwei (¾ * 0.5).