EVM Kernel: FA deposit v2
In the first iteration of the FA Bridge featured by the Bifrost kernel
proposal, the execution of the FA deposits is sponsored by the chain. This
design choice was made to ease the use of the FA bridge by third-parties, but
came with a set of constraints (most notably, on how much gas a FA deposit can
use).

This patch revisits this design choice to retire the gas sponsoring and align
the execution of FA deposits with regular transactions. To a large extent, it
is a logical next step for the FA bridge, akin to removing “training wheels.”

Prior to this patch, a FA deposit was directly executed when injected in a
block by the sequencer. This was manifested by a transaction from the system
address (`0x0000..00`) to the extended ERC-20 proxy contract, with a log to
advertise the transfer of ownership. For instance, the FA deposit
`0x6ce1063705c99cce563e127ccacc65fda5d005594246a6671d56613cba6e27b0` injected
on Etherlink mainnet emitted the following event.

    {
      "address": "0x0000000000000000000000000000000000000000",
      "topics": [
        "0x7ee7a1de9c18ce695c95b8b19fbdf26cce3544e3ca9e08c9f487776783d7599f",
        "0x37e070b9449b3f88dbf5a9f5de742d844cf0c4a14296087353ae06347fde824b"
      ],
      "data": "0x0000000000000000000000000d48d8079631068e95d8cff1ed62e9302404fbac000000000000000000000000961a4f3eb5b8876e01fdc708f88254e5b1b4d9bd0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000085a6c30000000000000000000000000000000000000000000000000000000000000010",
      "blockNumber": "0xc11695",
      "transactionHash": "0x6ce1063705c99cce563e127ccacc65fda5d005594246a6671d56613cba6e27b0",
      "transactionIndex": "0x0",
      "blockHash": "0x9c7c64faf893e8082b7797bfd7d2555768a1fc13733a1f80716c0da95334f9ec",
      "logIndex": "0x1",
      "removed": false
    }

With this patch, the worflow changes. When the sequencer injects the FA deposit
to the chain, instead of being executed, the deposit is simply put in a table.
On-chain, this manifests with a new transaction going from the system address
`0x0000..00` to the FA bridge precompile smart contract `0xff00..02`. This
transactions emits a new log containing notably a nonce associated with the
deposit. See the event emitted by the same FA deposit, this time run using this
kernel.

    {
      "address": "0xff00000000000000000000000000000000000002",
      "topics": [
        "0x27a88c034649434b9c0dd0bf36ed46822cad6427dab69d15870da95c0e069acd",
        "0x37e070b9449b3f88dbf5a9f5de742d844cf0c4a14296087353ae06347fde824b",
        "0x0000000000000000000000000d48d8079631068e95d8cff1ed62e9302404fbac"
      ],
      "data": "0x0000000000000000000000000000000000000000000000000000000000000258000000000000000000000000961a4f3eb5b8876e01fdc708f88254e5b1b4d9bd0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000085a6c30000000000000000000000000000000000000000000000000000000000000010",
      "blockNumber": "0xc11695",
      "transactionHash": "0x6ce1063705c99cce563e127ccacc65fda5d005594246a6671d56613cba6e27b0",
      "transactionIndex": "0x0",
      "blockHash": "0x782b28a706efed42cca188e772289775aa9a8a76c6a1e80c3df8b1ee358879be",
      "logIndex": "0x0",
      "removed": false
    }

About the topics, in addition to the selector,

  - `0x37e070b9449b3f88dbf5a9f5de742d844cf0c4a14296087353ae06347fde824b` is the
    ticket hash.
  - The lower half of `0x00..00d48d8079631068e95d8cff1ed62e9302404fbac` is the
    proxy contract address.

The first 32 bytes of `data` give the nonce associated to the deposit (0x258,
that is 600).

Once the FA deposit is “queued” (that is, injected by the sequencer), it
becomes possible to claim it with a new entrypoint of the FA bridge precompile
contract, `claim(uint256)`.

Anyone can inject a claim transaction, and pay for the gas necessary to execute
this transaction. Executing a `claim` leads to the initial event being emitted.
As a consequence, pre-existing deployments should work as expected, as long as
a new component is deployed to execute claims.

Co-authored-by:
Thomas Letan <lthms@nomadic-labs.com>