EVM Kernel: Update governance contracts of D
In Rio, the cycle duration has changed, which has an impact on the governance
of Tezos. Transitively, this has an impact of the governance of Etherlink,
which relies on the voting power of bakers.

We have deployed new governance contracts accordingly, and this patch
introduces the necessary migration for Etherlink to start using them.