Kernel/Sdk: Stop force encoding to Path when deriving an enum
When deriving NomReader, BinWriter or HasEncoding, we build an encoding that will be our
guideline to use the right function. But when building the encoding of an enum, we force
the encoding to be a Path. It means that the type in the enum must explicitly implement
the trait we are trying to derive.
For example, a 'Vec<u8>' is a good field for a struct when deriving NomReader, but when
putting 'Vec<u8>' in an enum, it doesn't work anymore and we must explicitly implement
NomReader for 'Vec<u8>'