L2/node: Introduce a generic block number rpc and use it in blueprint follower
Blueprint follower was using `eth_blockNumber` to follow the chain. But in Tezlink
mode `eth_blockNumber` is returning an error because searching in `/evm/world_state`.
So we introduce a new rpc or the sequencer that will return the block number depending
on the chain family.

The tezt that verify that restricted rpcs worked was modified because it was disabling
`tez_blockNumber` and make the node crash