CI: create scheduled pipeline for building documentation
Contains job for static x86_64 binaries so we can reuse directly documentation jobs from master_branch.ml.

It should be ok as it is a scheduled pipeline.

We may be able to shorten this pipeline by depending on jobs building non-static binaries, but that would I think add quite a few jobs.
Plus some refactoring might be needed.