EVM node/Sequencer: assert that there is only one chain
Until now, we avoided to assert that the sequencer was handling a
single chain to avoid going against future multichain support. This
lead us in an hybrid system in which some workers expect a single
chain (and a single tx container) and some other don't. Assuming
everywhere that there is, for now, only one chain does however
simplify the refactoring code needed to plug the tx queue in Tezlink
so in this commit we finally byte this bullet.