EVM/Node: fix [accessList] field
As of right now, when the transaction is an EIP-1559
one, we omit the accessList from the outputed transaction
object if it's empty.
That's NOT standard.
The commit contain two fix in one, we return the access list
wether it's empty or not, and the field is renamed to [accessList]
and not [access_list].