DAL: simplify function cells_of_level's implementation
Starting from Rio, the function `cells_of_level` exposed by the DAL plugin
can work with just an attestation level as an argument instead of
block_info.

This MR proposes to make the simplification. For Quebec, we (re)-fetch the
block info locally, as its metadata is needed. This should not be an
issue in practice, as Rio has been activated for almost one month
(Quebec's `cells_of_level` function will likely not be called).