EVM Node: Add `tez_getTransactionGasInfo` RPC
* What

This commit introduces a new Etherlink-specific RPC method,
`tez_getTransactionGasInfo`. This RPC enables users to retrieve the exact split
between execution and inclusion gas of included transactions.

* Why

The standard Ethereum `eth_getTransactionReceipt` RPC provides `gasUsed`, which
in the case of Etherlink includes both EVM execution gas and inclusion fees.
However, it is essential to distinguish the gas spent solely on EVM execution,
separate from inclusion costs costs. This new RPC fulfills that specific
requirement.

Even if not used by our users, it can be a useful tool for troubleshooting. Not
being able to distinguish between inclusion fees and execution fees have been
an issue in the past.

* How

The `tez_getTransactionGasInfo` RPC accepts a transaction hash. It first
retrieves the transaction receipt and object. Subsequently, it fetches the
minimum base fee per gas and the inclusion fee per byte from the durable
storage. The inclusion fees are then calculated using the transaction's
calldata and subtracted from the total `gasUsed` found in the transaction
receipt, resulting in the pure execution gas.