EVM Node: Advertises execution gas on blueprint application
* What

This commit modifies the blueprint application event to include
`execution_gas`. This new metric represents the gas consumed by the
transactions within a blueprint, excluding the gas used for inclusion fees.

* Why

The addition of `execution_gas` provides a more precise measure of transaction
processing cost. This distinction is crucial for monitoring the efficiency of
the EVM and for separating the actual computation cost from the overhead
associated with publishing data to the L1.

* How

The `blueprint_application` event declaration in `blueprint_events.ml` has been
updated to expect an additional `execution_gas` parameter. The
`blueprint_applied` function now passes this new metric when emitting the
event. The calculation for `execution_gas` is performed within
`evm_context.ml`, by subtracting the inclusion fees from the `gas_used`
reported in the transaction receipt, for each transaction.