tezt/evm: register event handler in run and not create
The declaration of the event handler was done in `Evm_node.create`,
but the cleaning in case of terminate was done in `Evm_node.run`. IT's
clearer to register both closer, also it fixes an issue when we
restart a node using `Evm_node.terminate` then `Evm_node.run` where
the `wait_for_ready` was timeouted because the event handler was not
defined anymore.