Etherlink/Revm: minimal error for REVM
The goals here is to keep the error interface as
minimal as possible. No need to be super-extensive.

In most case either the transaction succeed or we're
in an unrecoverable state.

In the second case it's either a Runtime error or
what I've called a custom error which we don't necessarily
need to know what it is at the enum level.
As long as it's logged we get the info of what happened and
that's it. No need to over-engineer this part and
make the error handling to complicated.

If changes are required, we can revisit later.