EVM Kernel: Refactor speed limit calculation
* What

This commit renames the gas price parameter `speed_limit` to `target`
in both the OCaml and Rust implementations of the gas price module.
The `gas_constants` in the OCaml code are also updated to derive the
`target` from a `capacity` value, specifically setting it to half of
the defined capacity.

* Why

The change to `target` from `speed_limit` provides a more accurate
and descriptive representation of the desired average gas consumption
rate that the system aims to achieve. This aligns better with common
gas pricing mechanisms, which often aim for a target block utilization
rather than a strict speed limit.

* How

The constants in `gas_price.rs` and corresponding type in `gas_price.ml` have
been updated. All references to `speed_limit` were replaced with `target`. A
new `capacity` constant was introduced in OCaml's `gas_constants` to calculate
the `target` value.